daylight_saving func of proper type and cron_func_t used in cron table instead.

+++ src/clock.c
@@ -57,7 +57,7 @@ uint8_t month_len()
   return len;
 }
 
-void daylight_saving(uint16_t hours)
+void daylight_saving(int8_t hours)
 {
   date.hour += (int8_t)hours;
 }
